+++ server/routes/category.js
@@ -1,2 +1,10 @@
 // routeto one category
 // will have data of all the playlists under the category
+
+const express = require('express');
+const router = express.Router();
+
+module.exports = (db) => {
+
+  return router
+};
