+++ server/routes/dashboard.js
@@ -1,2 +1,10 @@
 // route to a playlist dashbord
 // will have all the stats for the selected playlist
+
+const express = require('express');
+const router = express.Router();
+
+module.exports = (db) => {
+
+  return router
+};
