+++ server/routes/index.js
@@ -1,2 +1,10 @@
 // route to home page
 // will display all the categories for that user
+
+const express = require('express');
+const router = express.Router();
+
+module.exports = (db) => {
+
+  return router
+};
